requirements.txt
asgiref==3.7.2
Django==5.0
psycopg2-binary==2.9.9
sqlparse==0.4.4
tzdata==2023.3
gunicorn==21.2.0
python-dotenv==1.0.0
whitenoise==6.6.0
pip==23.3.1
djangorestframework~=3.14.0